Revert "Elastic Integration (Phase II)"
