+++ lib/bootstrap.js
@@ -264,10 +264,7 @@ module.exports = function () {
         log('error', 'bootstrap error:', ex);
       });
     });
-  })).return(null);
-  // Return null because no values from the boostrapping
-  // are used in the next instantiation steps, but returning
-  // a value here affects how we connect to Elastic Search
+  }));
 };
 
 module.exports.bootstrapPath = bootstrapPath;
