--- lib/elastic/mappings/pages.yml
@@ -1,46 +1,0 @@
-general:
-  dynamic: strict
-  properties:
-    title:
-      type: string
-      index: not_analyzed
-    published:
-      type: boolean
-      index: not_analyzed
-    scheduled:
-      type: boolean
-      index: not_analyzed
-    url:
-      type: string
-      index: not_analyzed
-    uri:
-      type: string
-      index: not_analyzed
-    site:
-      dynamic: strict
-      type: object
-      properties:
-        name:
-          type: string
-          index: not_analyzed
-        slug:
-          type: string
-          index: not_analyzed
-        host:
-          type: string
-          index: not_analyzed
-        path:
-          type: string
-          index: not_analyzed
-        port:
-          type: string
-          index: not_analyzed
-        assetDir:
-          type: string
-          index: not_analyzed
-        assetPath:
-          type: string
-          index: not_analyzed
-  _timestamp:
-    enabled: true
-    store: yes
