+++ lib/files.js
@@ -9,7 +9,6 @@
 'use strict';
 
 let _ = require('lodash'),
-  bluebird = require('bluebird'),
   control = require('./control'),
   fs = require('fs'),
   path = require('path'),
@@ -285,40 +284,12 @@ function getComponentPackage(name) {
  * @return {Promise}
  */
 function readFilePromise(file) {
-  return new bluebird((resolve, reject) => {
+  return new Promise((resolve, reject) => {
     fs.readFile(file, 'utf8', (err, data) => {
       if (err) {
         reject(err);
       }
       resolve(data);
-    });
-  });
-}
-
-/**
- * Returns a promise representing the retrieval of all files
- * in a directory
- *
- * @param {string} dir
- * @returns {Promise}
- */
-function getFilesPromise(dir) {
-  return new bluebird((resolve, reject) => {
-    fs.readdir(dir, function (err, data) {
-      var filteredFiles;
-
-      if (err) {
-        reject(err);
-      }
-
-      // Filter out certain files and directories
-      filteredFiles = _.filter(data, (file) => {
-        return !exports.isDirectory(path.join(dir, file)) &&
-          file.indexOf('.test.') === -1 && // ignore test files
-          file.indexOf('.md') === -1; // ignore documentation files
-      });
-
-      resolve(filteredFiles);
     });
   });
 }
@@ -334,11 +305,9 @@ exports.getComponentPackage = control.memoize(getComponentPackage);
 exports.isDirectory = control.memoize(isDirectory);
 exports.tryRequire = control.memoize(tryRequire);
 exports.readFilePromise = control.memoize(readFilePromise);
-exports.getFilesPromise = control.memoize(getFilesPromise);
 
 // for testing
 exports.setPackageConfiguration = setPackageConfiguration;
 exports.setRequire = setRequire;
 exports.setAllowedEnvFiles = setAllowedEnvFiles;
 exports.getAllowedEnvFiles = getAllowedEnvFiles;
-
