+++ lib/files.test.js
@@ -364,31 +364,4 @@ describe(_.startCase(filename), function () {
       });
     });
   });
-
-  describe('getFilesPromise', function () {
-    const fn = lib[this.title];
-
-    it('gets a list of folders', function () {
-      sandbox.stub(fs, 'readdir', function (path, callback) {
-        return callback(null, ['isAFile', 'isADirectory']);
-      });
-      fs.statSync.withArgs('isAFile').returns(createMockStat({isDirectory: false}));
-      fs.statSync.withArgs('isADirectory').returns(createMockStat({isDirectory: true}));
-      path.resolve.returnsArg(0);
-
-      fn('.').then(function (result) {
-        expect(result).to.contain('isAFile');
-      });
-    });
-
-    it('returns empty array on error', function () {
-      sandbox.stub(fs, 'readdir', function (path, callback) {
-        return callback(new Error('Error'), '');
-      });
-
-      fn('test').catch(function (result) {
-        expect(result).to.equal('');
-      });
-    });
-  });
 });
