+++ lib/services/search.js
@@ -3,10 +3,6 @@
 const elastic = require('elasticsearch'),
   bluebird = require('bluebird'),
   _ = require('lodash'),
-  elasticHelpers = require('../elastic/helpers'),
-  files = require('../files'),
-  moduleExists = require('../utils/module-exists'),
-  path = require('path'),
   util = require('util'),
   log = require('./log').withStandardPrefix(__filename),
   endpoint = process.env.ELASTIC_HOST || '',
@@ -16,12 +12,10 @@ const elastic = require('elasticsearch'),
     apiVersion: '1.7',
     // log: ['error', 'trace'],
     defer: require('../utils/defer')
-  },
-  searchDir = '/search';
+  };
 
 // Reference to the ES client
 var client;
-
 
 /**
  * Test whether or not the client is connected
@@ -225,6 +219,38 @@ function createAliasIfNone(index) {
     });
 }
 
+/**
+ * Convert Redis batch operations to Elasticsearch batch operations
+ *
+ * @param {string} index
+ * @param {string} type
+ * @param {Array} ops
+ * @returns {Array}
+ */
+function convertRedisBatchtoElasticBatch(index, type, ops) {
+  let bulkOps = [];
+
+  _.each(ops, function (op) {
+    if (_.isString(op.value)) {
+      throw new TypeError('op.value cannot be string: ' + JSON.stringify(op));
+    } else if (op.type === 'put') {
+      let indexOp = {_index: index, _type: type};
+
+      // key is optional; if missing, an id will be generated that is unique across all shards
+      if (op.key) {
+        indexOp._id = op.key;
+      }
+
+      bulkOps.push(
+        {index: indexOp},
+        op.value
+      );
+    } else {
+      log('warn', 'Unhandled batch operation:',  op);
+    }
+  });
+  return bulkOps;
+}
 
 /**
  * Query an Elasticsearch index
@@ -313,62 +339,17 @@ function batch(ops) {
 }
 
 /**
- * Create the ES Client or set the client to the overrideClient
- * @param  {Object} [overrideClient]
- * @return {Promise}
- */
-function createClient(overrideClient) {
-  return new bluebird(function (resolve, reject) {
-    if (!module.exports.endpoint && !overrideClient) {
-      reject(new Error('No Elastic endpoint or client override'));
-    }
-
-    // Set the exported client
-    module.exports.client = client = module.exports.endpoint && !overrideClient ? new elastic.Client(_.clone(serverConfig)) : _.assign(module.exports.client, overrideClient);
-
-    resolve(client);
-  });
-}
-
-/**
- * Setup the client and check the health of the client. Override
- * client is an optional argument which will replace the regular
- * ES client. This is used for testing, but in production
- * `overrideClient` will be undefined.
+ * Create the ES Client or an empty object
  *
- * @param {Object} [overrideClient]
- * @return {Promise}
+ * @param {Object} overrideClient
  */
 function setup(overrideClient) {
-  // Create client + check the connection
-  return createClient(overrideClient)
-    .then(healthCheck)
-    // .then(createInternalIndices)
-    .then(setupMappings)
-    .catch(function (err) {
-      return bluebird.reject(err);
-    });
-}
+  if (!module.exports.endpoint && !overrideClient) {
+    throw new Error('No Elastic endpoint or client override');
+  }
 
-/**
- * Check if the correct indices exist, and if they don't, create them.
- *
- * @param {Object} mappings
- * @returns {Promise}
- */
-function validateIndices(mappings) {
-  return bluebird.all(_.map(Object.keys(mappings), function (index) {
-    return module.exports.createIndexIfNone(createIndexName(index)).then(function () {
-      return module.exports.createAliasIfNone(index);
-    });
-  }))
-  .then(function () {
-    return bluebird.all(_.reduce(mappings, function (list, types, index) {
-      return list.concat(_.map(types, function (mapping, type) {
-        return module.exports.createMappingIfNone(createIndexName(index), type, mapping);
-      }));
-    }, []));
-  });
+  // Set the exported client
+  module.exports.client = client = module.exports.endpoint && !overrideClient ? new elastic.Client(_.clone(serverConfig)) : _.assign(module.exports.client, overrideClient);
 }
 
 /**
@@ -377,78 +358,11 @@ function validateIndices(mappings) {
  * endpoint. User is responsible for sending
  * the proper payload
  *
- * @param {object} queryObj
+ * @param {object} queryObj [description]
  * @returns {Promise}
  */
 function RESTQuery(queryObj) {
   return client.search(queryObj);
-}
-
-/**
- * Load the mappings from the mappings folder.
- *
- * The name of the file becomes the name of the index (convention over configuration).
- *
- * @param {string} dirPath
- * @returns {object}
- */
-function retrieveMappingData(dirPath) {
-  var list = files.getFiles(dirPath),
-    mappings = {};
-
-  _.each(list, function (filename) {
-    const mappingName = filename.split('.')[0],
-      mapping = files.getYaml(path.join(dirPath, mappingName));
-
-    mappings[mappingName] = mapping;
-  });
-
-  return mappings;
-}
-
-// /**
-//  * Create indices that are required for Clay
-//  *
-//  * @return {Promise}
-//  */
-// function createInternalIndices() {
-//   var dirPath = path.resolve(__dirname, '../elastic/mappings'),
-//     internalMappings = retrieveMappingData(dirPath);
-//
-//   return validateIndices(internalMappings)
-//     .then(require(process.cwd() + searchDir))
-//     .then(function () {
-//       // Listen to events internally
-//       db.on('batch', pageList.updatePageList(internalMappings));
-//       return bluebird.resolve();
-//     });
-// }
-
-/**
- * If a directory with search data exists, instantiate it.
- * If not, don't do anything.
- *
- * @param {String} overrideDir
- * @return {Promise}
- */
-function setupMappings(overrideDir) {
-  var externalSearch = moduleExists(process.cwd() + (overrideDir || searchDir));
-
-  if (externalSearch) {
-    return externalSearch();
-  }
-
-  return bluebird.resolve();
-}
-
-/**
- * Return current instance of the search client. Also
- * available via the `module.exports.client`.
- *
- * @return {object}
- */
-function getInstance() {
-  return module.exports.client;
 }
 
 module.exports.setup = setup;
@@ -463,25 +377,11 @@ module.exports.existsIndex = existsIndex;
 module.exports.initMapping = initMapping;
 module.exports.existsMapping = existsMapping;
 module.exports.createMappingIfNone = createMappingIfNone;
+module.exports.convertRedisBatchtoElasticBatch = convertRedisBatchtoElasticBatch;
 module.exports.del = del;
 module.exports.batch = batch;
 module.exports.put = put;
 module.exports.query = query;
 module.exports.RESTQuery = RESTQuery;
-module.exports.retrieveMappingData = retrieveMappingData;
-module.exports.validateIndices = validateIndices;
-
-// Filters for Elastic ops
-module.exports.filters = require('../elastic/db-filters');
-// Elastic Helpers
-module.exports.convertRedisBatchtoElasticBatch = elasticHelpers.convertRedisBatchtoElasticBatch;
-module.exports.normalizeOpValuesWithMapping = elasticHelpers.normalizeOpValuesWithMapping;
-module.exports.parseOpValue = elasticHelpers.parseOpValue;
-module.exports.applyOpFilters = elasticHelpers.applyOpFilters;
-module.exports.removeAllReferences = elasticHelpers.removeAllReferences;
-module.exports.getInstance = getInstance;
-
 // Exported for testing
 module.exports.createIndexName = createIndexName;
-module.exports.createClient = createClient;
-module.exports.setupMappings = setupMappings;
