+++ lib/services/search.test.js
@@ -2,8 +2,6 @@
 /* eslint max-nested-callbacks:[2,5] */
 
 var _ = require('lodash'),
-  bluebird = require('bluebird'),
-  files = require('../files'),
   sinon = require('sinon'),
   expect = require('chai').expect,
   filename = __filename.split('/').pop().split('.').shift(),
@@ -48,13 +46,14 @@ describe(_.startCase(filename), function () {
     sandbox.restore();
   });
 
-  describe('createClient', function () {
-    const fn = lib[this.title];
-
+  // setup
+  describe('setup', function () {
     it('creates an Elastic client if an endpoint is defined and no override is passed in', function () {
-      fn().catch(function (err) {
-        expect(err).to.be.an.instanceof(Error);
-      });
+      var result = function () {
+        lib.setup();
+      };
+
+      expect(result).to.throw(Error);
     });
 
     it('returns an instance with fake client if one is passed in', function () {
@@ -69,6 +68,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // healthCheck
   describe('healthCheck', function () {
     var fn = lib[this.title];
 
@@ -87,6 +87,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // existsMapping
   describe('existsMapping', function () {
     var fn = lib[this.title];
 
@@ -99,6 +100,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // initIndex
   describe('initIndex', function () {
     var fn = lib[this.title];
 
@@ -119,6 +121,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // initAlias
   describe('initAlias', function () {
     var fn = lib[this.title];
 
@@ -139,6 +142,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // initMapping
   describe('initMapping', function () {
     var fn = lib[this.title];
 
@@ -159,6 +163,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // createAliasIfNone
   describe('createAliasIfNone', function () {
     var fn = lib[this.title];
 
@@ -195,6 +200,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // createMappingIfNone
   describe('createMappingIfNone', function () {
     var fn = lib[this.title];
 
@@ -226,6 +232,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // createIndexName
   describe('createIndexName', function () {
     var fn = lib[this.title];
 
@@ -234,6 +241,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // createIndexIfNone
   describe('createIndexIfNone', function () {
     var fn = lib[this.title];
 
@@ -270,6 +278,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // Exists Index
   describe('existsIndex', function () {
     var fn = lib[this.title];
 
@@ -282,6 +291,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // Exists Alias
   describe('existsAlias', function () {
     var fn = lib[this.title];
 
@@ -322,6 +332,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // Delete
   describe('del', function () {
     var fn = lib[this.title];
 
@@ -350,6 +361,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // Put
   describe('put', function () {
     var fn = lib[this.title];
 
@@ -378,6 +390,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // convertRedisBatchtoElasticBatch
   describe('convertRedisBatchtoElasticBatch', function () {
     var fn = lib[this.title];
 
@@ -387,7 +400,7 @@ describe(_.startCase(filename), function () {
           fn('index', 'type', ops);
         };
 
-      expect(result).to.throw(SyntaxError);
+      expect(result).to.throw(TypeError);
     });
 
     it('returns an array of ops if type property equals "put"', function () {
@@ -409,6 +422,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // batch
   describe('batch', function () {
     var fn = lib[this.title];
 
@@ -429,6 +443,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // batch
   describe('batch', function () {
     var fn = lib[this.title];
 
@@ -449,6 +464,7 @@ describe(_.startCase(filename), function () {
     });
   });
 
+  // RESTQuery
   describe('RESTQuery', function () {
     var fn = lib[this.title];
 
@@ -460,93 +476,5 @@ describe(_.startCase(filename), function () {
         });
     });
   });
-
-  describe('validateIndices', function () {
-    var fn = lib[this.title],
-      mapping = {
-        pages: {
-          general: {
-            dynamic: 'strict',
-            properties: {
-              prop: { type: 'string', index: 'not_analyzed' },
-            }
-          }
-        }
-      };
-
-    it('calls createIndexIfNone, createAliasIfNone and createMappingIfNone', function () {
-      sandbox.stub(lib, 'createIndexIfNone');
-      sandbox.stub(lib, 'createAliasIfNone');
-      sandbox.stub(lib, 'createMappingIfNone');
-      lib.createIndexIfNone.returns(bluebird.resolve('index'));
-      lib.createAliasIfNone.returns(bluebird.resolve('alias'));
-      lib.createMappingIfNone.returns(bluebird.resolve('mapping'));
-
-      return fn(mapping)
-        .then(function () {
-          sinon.assert.calledOnce(lib.createIndexIfNone);
-          sinon.assert.calledOnce(lib.createAliasIfNone);
-          sinon.assert.calledOnce(lib.createMappingIfNone);
-        });
-    });
-  });
-
-  describe('getInstance', function () {
-    var fn = lib[this.title];
-
-    it('returns the client', function () {
-      expect(fn()).to.deep.equal(lib.client);
-    });
-  });
-
-  describe('retrieveMappingData', function () {
-    var fn = lib[this.title];
-
-    it('returns a mapping object if the file exists', function () {
-      var result;
-
-      sandbox.stub(files, 'getFiles');
-      sandbox.stub(files, 'getYaml');
-      files.getFiles.returns(['pages.yml']);
-      files.getYaml.returns({ pages: { mapping: true } });
-      result = fn('path');
-      expect(result).to.deep.equal({ pages: { pages: { mapping: true } } });
-    });
-
-    it('returns an empty object if no mapping exists', function () {
-      var noFilesResult;
-
-      sandbox.stub(files, 'getFiles');
-      sandbox.stub(files, 'getYaml');
-      files.getFiles.returns([]);
-      files.getYaml.returns(null);
-      noFilesResult = fn('path');
-      expect(noFilesResult).to.deep.equal({});
-    });
-  });
-
-  describe('setupMappings', function () {
-    var fn = lib[this.title];
-
-    it('requires a module if it exists', function () {
-      return fn('/test/fixtures/module-exists')
-        .then(function (resp) {
-          expect(resp).to.equal('exists');
-        });
-    });
-
-    it('resolves with no response if the module does not exist', function () {
-      return fn('fake/path/to/module')
-        .then(function (resp) {
-          expect(resp).to.be.undefined;
-        });
-    });
-
-    it('uses the pre-defined search directory if none is provided', function () {
-      return fn()
-        .then(function (resp) {
-          expect(resp).to.be.undefined;
-        });
-    });
-  });
 });
+
