+++ lib/setup.js
@@ -4,8 +4,7 @@ const bluebird = require('bluebird'),
   express = require('express'),
   routes = require('./routes'),
   bootstrap = require('./bootstrap'),
-  htmlComposer = require('./html-composer'),
-  search = require('./services/search');
+  htmlComposer = require('./html-composer');
 
 bluebird.config({
   longStackTraces: true
@@ -38,7 +37,8 @@ module.exports = function (options) {
     htmlComposer.addEngines(engines);
   }
 
-  return bootstrap() // look for bootstraps in components
-    .then(search.setup) // setup search
-    .return(router);
+  // look for bootstraps in components
+  return bootstrap().then(function () {
+    return router;
+  });
 };
