+++ lib/setup.test.js
@@ -4,8 +4,7 @@ const _ = require('lodash'),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require('./' + filename),
   sinon = require('sinon'),
-  htmlComposer = require('./html-composer'),
-  search = require('./services/search');
+  htmlComposer = require('./html-composer');
 
 describe(_.startCase(filename), function () {
   let sandbox;
@@ -13,7 +12,6 @@ describe(_.startCase(filename), function () {
   beforeEach(function () {
     sandbox = sinon.sandbox.create();
     sandbox.stub(htmlComposer, 'addEngines');
-    sandbox.stub(search, 'setup');
   });
 
   afterEach(function () {
@@ -26,7 +24,6 @@ describe(_.startCase(filename), function () {
 
   it('adds engines', function () {
     return lib({engines: 'whatever'}).then(function () {
-      sinon.assert.calledOnce(search.setup);
       sinon.assert.calledOnce(htmlComposer.addEngines);
     });
   });
