--- lib/utils/module-exists.js
@@ -1,11 +1,0 @@
-'use strict';
-
-module.exports = function (modulePath) {
-  try {
-    require(modulePath);
-  } catch (e) {
-    return;
-  }
-
-  return require(modulePath);
-};
