+++ package.json
@@ -32,7 +32,6 @@
     "js-yaml": "^3.2.0",
     "levelup": "^1.3.0",
     "lodash": "^4.14.2",
-    "lodash-ny-util": "^0.1.1",
     "memdown": "^1.0.0",
     "multiplex-templates": "^1.2.0",
     "node-fetch": "^1.3.0",
