--- test/fixtures/module-exists.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const bluebird = require('bluebird');
-
-module.exports = function () {
-  return bluebird.resolve('exists');
-};
